Reorganized project structure
