+++ src/Lists/DoublyLinkedIntList.java
@@ -1,16 +1,39 @@
+package Lists;
+
 import java.util.Iterator;
+import Interfaces.IntList;
 
-public class ArrayIntList implements IntList{
+public class DoublyLinkedIntList implements IntList {
 
-    // fields:
+    // Fields
+    private Node left;
+    private Node right;
     private int size;
-    private int[] buffer;
 
-    public ArrayIntList() {
-        //initialize fields
+    // Constructor
+    public DoublyLinkedIntList() {
+        // an empty list has two sentinel (dummy) nodes that serve as bookends
+        left = new Node();
+        right = new Node();
+        left.next = right;
+        right.prev = left;
         size = 0;
-        buffer = new int[10];
+
     }
+
+    private class Node {
+        int data;
+        Node next;  // address of the node 'after' this one in line
+        Node prev;  // addres of the node 'before' this one in line
+
+        public Node() {
+            next = null;
+            prev = null;
+        }
+    }
+
+
+
     /**
      * Prepends (inserts) the specified value at the front of the list (at index 0).
      * Shifts the value currently at the front of the list (if any) and any
@@ -20,14 +43,6 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public void addFront(int value) {
-        for (int i = size; i >= 0; i--) {
-            buffer[i] = buffer[i - 1];
-
-        }
-
-        // put the value at the front of the array at position 0
-        buffer[0] = value;
-        size++;
 
     }
 
@@ -38,13 +53,20 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public void addBack(int value) {
-        //TODO: check to see if we are full - if so, we need to create a larger buffer
+        Node theLastNode = right.prev;
 
-        if ( size == buffer.length) {
-            resize(size * 2);
-        }
+        // set up my new node and fill it out (data, prev, next)
+        Node theNewNode = new Node();
+        theNewNode.data = value;
+        theNewNode.next = right;
+        theNewNode.prev = theLastNode;
 
-        buffer[size] = value;
+        // go to the end of the list's sentinel and update it's prev
+        right.prev = theNewNode;
+
+        // go to the node before the new one and update it's next
+        theLastNode.next = theNewNode;
+
         size++;
 
     }
@@ -60,9 +82,6 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public void add(int index, int value) {
-        if ( size == buffer.length) {
-            resize(size * 2);
-        }
 
     }
 
@@ -82,7 +101,20 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public void removeBack() {
+        if (size > 0) {
+            // set up a temp variable for convenience
+            Node theOneToRemove = right.prev;
 
+            theOneToRemove.prev.next = right;
+            right.prev = theOneToRemove.prev;
+
+            // optional to clean up
+            theOneToRemove.next = null;
+            theOneToRemove.prev = null;
+            theOneToRemove.data = 0;
+
+            size--;
+        }
     }
 
     /**
@@ -96,26 +128,7 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public int remove(int index) {
-        // first, check the index to see if it is valid
-        if ( index < 0) {
-            throw new IndexOutOfBoundsException("Index cannot be negative");
-        } else if (index >= size) {
-            throw new IndexOutOfBoundsException("Index is higher than size");
-        }
-
-        // save a copy of the value to be removed so we can return it later
-        int copyOfRemovedValue = buffer[index];
-
-        //shift values to the left
-        for (int i = 0; i <= size - 1; i++) {
-            buffer[i] = buffer[i + 1];
-
-        }
-
-        buffer[size - 1] = 0;
-        size--;
-        
-        return copyOfRemovedValue;
+        return 0;
     }
 
     /**
@@ -127,10 +140,7 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public int get(int index) {
-        if (index > size) {
-            throw new IndexOutOfBoundsException("No Such Index Value...");
-        }
-        return buffer[index];
+        return 0;
     }
 
     /**
@@ -183,31 +193,7 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public void clear() {
-//        for (int i = 0; i < size; i++) {
-//            buffer[i] = 0;
-//        }
-//
-//        size = 0;
 
-        buffer = new int[10];
-        size = 0;
-    }
-
-    private void resize(int newSize) {
-        //create new space, separate from the old space (buffer)
-        int[] newBuffer = new int[newSize];
-
-        // copy everything over from buffer into newBuffer
-        for (int i = 0; i < buffer.length; i++) {
-            newBuffer[i] = buffer[i];
-
-        }
-
-        // set the new space into buffer
-        buffer = newBuffer;
-
-        // the old space is no longer "pointed to" and will eventually
-        // be cleaned up by the garbage collector
     }
 
     /**
@@ -217,50 +203,6 @@ public class ArrayIntList implements IntList{
      */
     @Override
     public Iterator<Integer> iterator() {
-
-        //iterators are what enables main/client to use a for-each lop on IntList
         return null;
     }
-
-    //create a private helper Iterator class
-    private class IntListIterator implements Iterator<Integer> {
-
-        // private fields:
-        private int i;
-
-        private IntListIterator() {
-            i = 0;
-        }
-
-        /**
-         * Returns {@code true} if the iteration has more elements.
-         * (In other words, returns {@code true} if {@link #next} would
-         * return an element rather than throwing an exception.)
-         *
-         * @return {@code true} if the iteration has more elements
-         */
-        @Override
-        public boolean hasNext() {
-            return i < size;
-        }
-
-        /**
-         * Returns the next element in the iteration.
-         *
-         * @return the next element in the iteration
-         * @throws NoSuchElementException if the iteration has no more elements
-         */
-        @Override
-        public Integer next() {
-            //check to see if i is greater than size
-//            if ( i >= size) {
-//                throw new
-//            }
-            
-            int currentValue = buffer[i];
-            i++;
-            return currentValue;
-        }
-    }
-
 }
