+++ src/driver/Driver.java
@@ -1,0 +1,16 @@
+package driver;
+
+import  Lists.DoublyLinkedIntList;
+
+
+
+public class Driver {
+
+    public static void main(String[] args) {
+        DoublyLinkedIntList firstList = new DoublyLinkedIntList();
+        firstList.addBack(0);
+
+        System.out.println(firstList);
+
+    }
+}
